requirements.txt
pandas
matplotlib.pyplot
BytesIO
tabula  # For reading PDF files